chore: align housekeeping and release workflow
