+++ .github/workflows/stable.yml
@@ -1,10 +1,12 @@
 name: Stable
 
 # Builds on every push to `stable`, ad-hoc signs, packages a DMG, publishes a `v<version>` release, and
-# regenerates this branch's appcast.xml (the stable channel's feed).
+# regenerates this branch's appcast.xml (the stable channel's feed), and merges the promoted stable
+# item into the nightly appcast so Nightly users can receive finished releases too.
 #
 # `stable` starts as a basic-files branch with no source — the build steps skip cleanly (green) until
-# code is promoted from `beta`/`nightly`. The stable version is the root VERSION file verbatim.
+# code is promoted from `beta`/`nightly`. The stable marketing version is the root VERSION file
+# verbatim; the build number is retained from the promoted nightly commit when possible.
 #
 # No Apple Developer ID / notarization. Only secret: SPARKLE_ED_PRIVATE_KEY (DMG still ships without it).
 
@@ -22,6 +24,7 @@ on:
       - 'LICENSE'
       - 'NOTICE'
       - '.gitignore'
+      - 'appcast.xml'
   workflow_dispatch:
 
 concurrency:
@@ -62,13 +65,16 @@ jobs:
       - name: Bundle (ad-hoc signed) + DMG
         if: steps.guard.outputs.build == 'true'
         run: |
-          export VERSION=$(cat VERSION 2>/dev/null || echo 1.0.0)
-          SHA=$(git rev-parse --short HEAD)
-          CHANNEL=stable ./scripts/bundle.sh release
+          git fetch origin nightly:refs/remotes/origin/nightly
+          SHA=$(./scripts/version-info.sh sha)
+          BUILD=$(BUILD_SOURCE_REF=origin/nightly SHA="${SHA}" ./scripts/version-info.sh build)
+          export BUILD SHA
+          export VERSION=$(CHANNEL=stable BUILD="${BUILD}" SHA="${SHA}" ./scripts/version-info.sh version)
+          CHANNEL=stable BUILD="${BUILD}" VERSION="${VERSION}" ./scripts/bundle.sh release
           mkdir -p updates
           brew install create-dmg
           ./scripts/make-dmg.sh stable Contained.app "updates/Contained-${VERSION}.dmg"
-          VERSION_VALUE="${VERSION}" ./scripts/release-body.sh > updates/release-notes.md
+          CHANNEL=stable VERSION_VALUE="${VERSION}" ./scripts/release-body.sh > updates/release-notes.md
           {
             echo "Contained ${VERSION} @ ${SHA}."
             echo
@@ -77,6 +83,7 @@ jobs:
             echo "⚠️ Unsigned build — first launch: right-click the app ▸ **Open** (or \`xattr -dr com.apple.quarantine /Applications/Contained.app\`)."
           } > updates/github-release-notes.md
           echo "VERSION=${VERSION}" >> "$GITHUB_ENV"
+          echo "BUILD=${BUILD}" >> "$GITHUB_ENV"
           echo "SHA=${SHA}" >> "$GITHUB_ENV"
 
       - name: Publish release
@@ -102,6 +109,28 @@ jobs:
           if git diff --cached --quiet; then
             echo "appcast unchanged — nothing to commit."
           else
-            git commit -m "ci: stable appcast (v${VERSION})"
+            git commit -m "ci: stable appcast (v${VERSION}) [skip ci]"
             git push origin HEAD:stable
           fi
+
+          PROMOTED_APPCAST="$(mktemp)"
+          cp appcast.xml "$PROMOTED_APPCAST"
+          git fetch origin nightly:refs/remotes/origin/nightly
+          NIGHTLY_WORKTREE="$(mktemp -d)"
+          rm -rf "$NIGHTLY_WORKTREE"
+          git worktree add "$NIGHTLY_WORKTREE" origin/nightly
+          ./scripts/promote-appcast-to-nightly.sh "$PROMOTED_APPCAST" "$NIGHTLY_WORKTREE/appcast.xml"
+          (
+            cd "$NIGHTLY_WORKTREE"
+            git config user.name  "github-actions[bot]"
+            git config user.email "github-actions[bot]@users.noreply.github.com"
+            git add appcast.xml
+            if git diff --cached --quiet; then
+              echo "nightly appcast already contains stable item."
+            else
+              git commit -m "ci: include stable appcast item (v${VERSION}) [skip ci]"
+              git push origin HEAD:nightly
+            fi
+          )
+          git worktree remove "$NIGHTLY_WORKTREE" --force
+          rm -f "$PROMOTED_APPCAST"
