+++ .gitignore
@@ -20,9 +20,17 @@ DerivedData/
 # App bundle (built artifact, regenerated by the bundling script)
 Contained.app/
 
-# Release artifacts (DMGs + the Sparkle staging dir). docs/appcast.xml IS committed (Pages-served).
+# Release artifacts (DMGs + Sparkle/release-note staging). Root appcast.xml IS committed on
+# release branches, but workflow path filters ignore appcast-only commits to avoid CI loops.
 *.dmg
 /updates/
+/.release/
+/.release-notes/
+
+# Change fragments are intended to be committed when used:
+#   changes/unreleased/*.md
+#   changes/nightly/*.md
+#   changes/beta/*.md
 
 # Captured CLI fixtures used only for local schema grounding (top-level only;
 # the committed test resources live under Tests/ContainedCoreTests/Fixtures).
